Update CalendarPage.jsx
Only the front end is implemented, and there is no back end support yet(refresh deleted events). Added ability to create event w/ event title, location, summary, hosting group, event coordinator, and coordinator's contact info. Updates the agenda page. When clicking on the event from the calendar, it opens the agenda page and autoscrolls to the selected event, which will be highlighted. Users can edit or delete events and the agenda page shows a range of events from the current date to the current date + 31 days.